Move core protocol implementation to protocols/core.rs
The switch function 'handle_request' is split between svsm_request.rs
(selecting the protocol) and protocols/core.rs (selecting the call).

No functional change intended.
